Move context commands to `.appium` ns.
