+++ lib/api/protocol/appium/getContext.js
@@ -1,4 +1,4 @@
-const ProtocolAction = require('./_base-action.js');
+const ProtocolAction = require('../_base-action.js');
 
 /**
  * Get the current context in which Appium is running. Used when testing hybrid mobile apps using Appium.
@@ -8,25 +8,26 @@ const ProtocolAction = require('./_base-action.js');
  * @example
  * module.exports = {
  *   'get current context': function (client) {
- *     client.getContext(function (result) {
- *       console.log('the current context is:', result.value);
- *     });
+ *     client
+ *       .appium.getContext(function (result) {
+ *         console.log('the current context is:', result.value);
+ *       });
  *   },
  *
  *   'get current context with ES6 async/await': async function (client) {
- *     const context = await client.getContext();
+ *     const context = await client.appium.getContext();
  *     console.log('the current context is:', context);
  *   }
  * };
  *
- * @syntax .getContext([callback])
+ * @syntax .appium.getContext([callback])
  * @method getContext
  * @param {function} [callback] Callback function which is called with the result value.
  * @returns {string|null} A string representing the current context, or `null` representing "no context".
  * @moreinfo appium.io/docs/en/writing-running-appium/web/hybrid/index.html
- * @see setContext
- * @see getContexts
- * @api protocol.mobile
+ * @see appium.setContext
+ * @see appium.getContexts
+ * @api protocol.appium
  */
 module.exports = class Session extends ProtocolAction {
   command(callback) {
