+++ lib/api/protocol/appium/getContexts.js
@@ -1,4 +1,4 @@
-const ProtocolAction = require('./_base-action.js');
+const ProtocolAction = require('../_base-action.js');
 
 /**
  * Get a list of the available contexts. Used when testing hybrid mobile apps using Appium.
@@ -8,25 +8,26 @@ const ProtocolAction = require('./_base-action.js');
  * @example
  * module.exports = {
  *   'get available contexts': function (client) {
- *     client.getContexts(function (result) {
- *       console.log('the available contexts are:', result.value);
- *     });
+ *     client
+ *       .appium.getContexts(function (result) {
+ *         console.log('the available contexts are:', result.value);
+ *       });
  *   },
  *
  *   'get available contexts with ES6 async/await': async function (client) {
- *     const contexts = await client.getContexts();
+ *     const contexts = await client.appium.getContexts();
  *     console.log('the available contexts are:', contexts);
  *   }
  * };
  *
- * @syntax .getContexts([callback])
+ * @syntax .appium.getContexts([callback])
  * @method getContexts
  * @param {function} [callback] Callback function which is called with the result value.
  * @returns {Array<string>} An array of strings representing available contexts, e.g 'WEBVIEW_<id>', or 'NATIVE_APP'
  * @moreinfo appium.io/docs/en/writing-running-appium/web/hybrid/index.html
- * @see getContext
- * @see setContext
- * @api protocol.mobile
+ * @see appium.getContext
+ * @see appium.setContext
+ * @api protocol.appium
  */
 module.exports = class Session extends ProtocolAction {
   command(callback) {
