+++ lib/api/protocol/contexts.js
@@ -9,7 +9,7 @@ const ProtocolAction = require('./_base-action.js');
  * @param {function} callback Callback function to be called when the command finishes.
  * @returns {Array} an array of strings representing available contexts, e.g 'WEBVIEW', or 'NATIVE'
  * @api protocol.mobile
- * @deprecated In favour of `.getContexts()`
+ * @deprecated In favour of `.appium.getContexts()`
  */
 module.exports = class Session extends ProtocolAction {
   command(callback) {
