+++ lib/api/protocol/currentContext.js
@@ -6,7 +6,7 @@ const ProtocolAction = require('./_base-action.js');
  * @param {function} callback Callback function to be called when the command finishes.
  * @returns {string|null} a string representing the current context or `null`, representing "no context"
  * @api protocol.mobile
- * @deprecated In favour of `.getContext()`
+ * @deprecated In favour of `.appium.getContext()`
  */
 module.exports = class Session extends ProtocolAction {
   command(callback) {
