+++ test/src/api/protocol/testContext.js
@@ -40,23 +40,37 @@ describe('context commands', function() {
     });
   });
 
-  it('testGetContexts', function() {
+  it('testGetContexts - appium', function() {
     return Globals.protocolTest({
       assertion: function(opts) {
         assert.strictEqual(opts.path, '/session/1352110219202/contexts');
       },
-      commandName: 'getContexts',
+      commandName: 'appium.getContexts',
       args: []
     });
   });
 
-  it('testGetContext', function() {
+  it('testGetContext - appium', function() {
     return Globals.protocolTest({
       assertion: function(opts) {
         assert.strictEqual(opts.path, '/session/1352110219202/context');
       },
-      commandName: 'getContext',
+      commandName: 'appium.getContext',
       args: []
+    });
+  });
+
+  it('testSetContext - appium', function() {
+    let text = 'NATIVE';
+
+    return Globals.protocolTest({
+      assertion: function(opts) {
+        assert.strictEqual(opts.method, 'POST');
+        assert.strictEqual(opts.path, '/session/1352110219202/context');
+        assert.deepStrictEqual(opts.data, {name: text});
+      },
+      commandName: 'appium.setContext',
+      args: [text]
     });
   });
 });
